OMG/run_instance/requirements.txt
google-cloud-compute==1.2.0
attrs==22.2.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
Fiona==1.9.2
google-api-core==2.11.0
google-auth==2.17.1
google-cloud-core==2.3.2
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
idna==3.4
munch==2.5.0
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyproj==3.5.0
requests==2.28.2
rsa==4.9
six==1.16.0
urllib3==1.26.15
functions-framework==3.3.0